parser_mask fix and formating

+++ components/sql/parser/base/parser_mask.cpp
@@ -1,8 +1,8 @@
 #include "parser_mask.hpp"
+#include <charconv>
 #include <memory_resource>
 #include <numeric>
 #include <set>
-#include <charconv>
 #include <string>
 
 namespace components::sql::impl {
@@ -129,8 +129,9 @@ namespace components::sql::impl {
             return document::value_t(tape, token_clean_value(token));
         } else if (token.type == token_type::number_literal) {
             if (is_integer(token.value())) {
-                int32_t result;
-                auto [ptr,ec]= std::from_chars(token.value().data(),token.value().data()+token.value().size(),result);
+                int64_t result;
+                auto [ptr, ec] =
+                    std::from_chars(token.value().data(), token.value().data() + token.value().size(), result);
                 return document::value_t(tape, result);
             } else {
                 // TODO we don't support float in this case?
